requirements.txt
ConfigArgParse==1.7
matplotlib==3.8.2
numpy==1.24.4
opencv_contrib_python==4.10.0.84
Pillow==10.1.0
Pillow==11.0.0
pytorch_lightning==2.1.2
scipy==1.14.1
skimage==0.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
